chore(cli): move 'chunk_path' to files.rs
